scada_fx5u_li/requirements.txt
pillow==12.2.0
pymcprotocol==0.3.0
opencv-python==4.13.0.92